Refactor filterpanel by splitting it up into components.
