+++ src/components/filterpanel/Header.js
@@ -1,0 +1,11 @@
+import React from "react";
+import logo from "../../hsl-logo.png";
+
+export default (props) => {
+  return (
+    <div>
+      <img src={logo} className="App-logo" alt="logo" />
+      <h1 className="App-title">Liikenteenvalvontatyökalu</h1>
+    </div>
+  );
+};
