+++ src/components/filterpanel/SuggestionInput.css
